requirements.txt
# Core Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# LangChain and AI
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langchain-core>=0.3.0

# LLM Providers
openai>=1.50.0
anthropic>=0.39.0

# Vector Store
faiss-cpu>=1.9.0

# Database (SQLite is built into Python - no additional packages needed)
# psycopg[binary]>=3.2.0  # OLD: PostgreSQL (removed for easier deployment)

# Utilities
python-dotenv>=1.0.0
requests>=2.32.0
pydantic>=2.9.0
pydantic-settings>=2.6.0

# Optional: For better performance
# langchain-anthropic>=0.3.0
# langchain-google-genai>=2.0.0

